Add task, user, and context type definitions
